[IMP] base: move `load_script` to odoo.modules.module
The function that directly imports a custom script file
into the namespace better belongs to the `module` module
rather than `migration` as we also use it to import
addons `commands`.
