+++ odoo/cli/command.py
@@ -7,8 +7,7 @@ from pathlib import Path
 
 import odoo.init  # import first for core setup
 import odoo.cli
-from odoo.modules import initialize_sys_path
-from odoo.modules.migration import load_script
+from odoo.modules import initialize_sys_path, load_script
 from odoo.tools import config
